requirements.txt

pandas>=2.2
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
prefect>=3.0
pytest>=8.0
